Create HtmlText Composable
